fix: Allow arbitrary permissions in v5
